refactor(api): restructure endpoints for better mobile development organization
Extract stats endpoints to separate /stats group and move recalculation
actions from /settings to /workouts for improved API organization.

Changes:
- Create new StatsEndpoints.cs with 8 stats endpoints
  - Move from /workouts/stats/* to /stats/*
  - Includes: weekly, yearly, relative-effort, best-efforts,
    yearly-weekly, available-periods, available-years
- Move recalculation endpoints from Settings to Workouts
  - Move from /settings/recalculate-* to /workouts/recalculate-*
  - Includes: relative-effort and splits recalculation (count + action)
- Update WorkoutsEndpoints.cs
  - Remove stats endpoint methods
  - Add recalculation endpoint methods
- Update SettingsEndpoints.cs
  - Remove recalculation methods (now in Workouts)
  - Keep only configuration endpoints
- Register StatsEndpoints in Program.cs
- Update frontend API client (12 endpoint URLs)
- Update Bruno test collection (12 test files)
- Update API reference documentation

BREAKING CHANGE: Endpoint URLs have changed:
- /workouts/stats/* → /stats/*
- /settings/recalculate-* → /workouts/recalculate-*

Functionality remains identical; only URL paths changed.
Frontend has been updated to use new paths.
