+++ api/bruno/Tempo.Api/Settings/Get count of workouts eligible for relative effort recalculation.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: {{baseUrl}}/settings/recalculate-relative-effort/count
+  url: {{baseUrl}}/workouts/recalculate-relative-effort/count
   body: none
   auth: inherit
 }
@@ -15,7 +15,7 @@ example {
   description: OK
   
   request: {
-    url: {{baseUrl}}/settings/recalculate-relative-effort/count
+    url: {{baseUrl}}/workouts/recalculate-relative-effort/count
     method: GET
     mode: none
   }
