+++ api/bruno/Tempo.Api/Settings/Get count of workouts eligible for split recalculation.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: {{baseUrl}}/settings/recalculate-splits/count
+  url: {{baseUrl}}/workouts/recalculate-splits/count
   body: none
   auth: inherit
 }
@@ -15,7 +15,7 @@ example {
   description: OK
   
   request: {
-    url: {{baseUrl}}/settings/recalculate-splits/count
+    url: {{baseUrl}}/workouts/recalculate-splits/count
     method: GET
     mode: none
   }
