+++ api/bruno/Tempo.Api/Settings/Recalculate relative effort for all qualifying workouts.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 post {
-  url: {{baseUrl}}/settings/recalculate-relative-effort
+  url: {{baseUrl}}/workouts/recalculate-relative-effort
   body: none
   auth: inherit
 }
@@ -15,7 +15,7 @@ example {
   description: OK
   
   request: {
-    url: {{baseUrl}}/settings/recalculate-relative-effort
+    url: {{baseUrl}}/workouts/recalculate-relative-effort
     method: POST
     mode: none
   }
@@ -40,7 +40,7 @@ example {
   description: Bad Request
   
   request: {
-    url: {{baseUrl}}/settings/recalculate-relative-effort
+    url: {{baseUrl}}/workouts/recalculate-relative-effort
     method: POST
     mode: none
   }
