+++ api/bruno/Tempo.Api/Workouts/Get available periods.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: {{baseUrl}}/workouts/stats/available-periods
+  url: {{baseUrl}}/stats/available-periods
   body: none
   auth: inherit
 }
@@ -19,7 +19,7 @@ example {
   description: OK
   
   request: {
-    url: {{baseUrl}}/workouts/stats/available-periods
+    url: {{baseUrl}}/stats/available-periods
     method: GET
     mode: none
     params:query: {
