+++ api/bruno/Tempo.Api/Workouts/Get available years.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: {{baseUrl}}/workouts/stats/available-years
+  url: {{baseUrl}}/stats/available-years
   body: none
   auth: inherit
 }
@@ -15,7 +15,7 @@ example {
   description: OK
   
   request: {
-    url: {{baseUrl}}/workouts/stats/available-years
+    url: {{baseUrl}}/stats/available-years
     method: GET
     mode: none
   }
