+++ api/bruno/Tempo.Api/Workouts/Get relative effort stats.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: {{baseUrl}}/workouts/stats/relative-effort
+  url: {{baseUrl}}/stats/relative-effort
   body: none
   auth: inherit
 }
@@ -19,7 +19,7 @@ example {
   description: OK
   
   request: {
-    url: {{baseUrl}}/workouts/stats/relative-effort
+    url: {{baseUrl}}/stats/relative-effort
     method: GET
     mode: none
     params:query: {
