+++ api/bruno/Tempo.Api/Workouts/Get weekly stats.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: {{baseUrl}}/workouts/stats/weekly
+  url: {{baseUrl}}/stats/weekly
   body: none
   auth: inherit
 }
@@ -19,7 +19,7 @@ example {
   description: OK
   
   request: {
-    url: {{baseUrl}}/workouts/stats/weekly
+    url: {{baseUrl}}/stats/weekly
     method: GET
     mode: none
     params:query: {
