+++ api/bruno/Tempo.Api/Workouts/Get yearly stats.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: {{baseUrl}}/workouts/stats/yearly
+  url: {{baseUrl}}/stats/yearly
   body: none
   auth: inherit
 }
@@ -19,7 +19,7 @@ example {
   description: OK
   
   request: {
-    url: {{baseUrl}}/workouts/stats/yearly
+    url: {{baseUrl}}/stats/yearly
     method: GET
     mode: none
     params:query: {
