+++ api/bruno/Tempo.Api/Workouts/Get yearly weekly stats.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: {{baseUrl}}/workouts/stats/yearly-weekly
+  url: {{baseUrl}}/stats/yearly-weekly
   body: none
   auth: inherit
 }
@@ -20,7 +20,7 @@ example {
   description: OK
   
   request: {
-    url: {{baseUrl}}/workouts/stats/yearly-weekly
+    url: {{baseUrl}}/stats/yearly-weekly
     method: GET
     mode: none
     params:query: {
